requirements.txt
streamlit>=1.32.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.24.0
pandas>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
urllib3>=2.0.0
